+++ src/Accounts/Accounts/ChangeLog.md
@@ -18,7 +18,7 @@
         - Additional information about change #1
 -->
 ## Upcoming Release
-* Fixed typo in comments
+* Fixed an issue that may cause Az to skip logs in Azure Automation or PowerShell jobs [#11492]
 
 ## Version 1.8.0
 * Updated `Add-AzEnvironment` and `Set-AzEnvironment` to accept parameters `AzureSynapseAnalyticsEndpointResourceId` and `AzureSynapseAnalyticsEndpointSuffix`